Create CoinFlip
Write a program that reads an arbitrary number of integers that are in the range of 0 to 100 inclusive.  This array will count how many occurrences of each are entered.  Indicate the end of input by a value outside the range.  Print all of the values (with the number of occurrences) that were entered zero or more times.